Desugar the label "*" into a unit transition.
In fact, the code for desugaring high-level topl labels into the
internal representation with (one or two) steps is completely
rewritten.  The rule is, roughly, that each half appears if either:
 - it was explicitly requested, or
 - there is some guard or action that implies it

This means that labels not containing "call" or "return" may result
in unit transitions, which is different from before. The * label is
just a special case of "there is a method name pattern, but no
call/return, nor any guard/action that identifies the time".