Go down the inheritance tree ignoring define methods.
For a method C.m() it will make a list with B.m() for *all* ancestors B
of D, regardless of whether they actually define a method m().  This
makes it possible to handle constructors in a more uniform way.

It should also be somewhat faster, at least for hard cases (many methods,
diamond inheritance).