Now toplc also compiles a main wrapper.
Previously, the wrapper tests/defaults/Config.java was compiled from the
script tests/defaults/run; now this is done by the '-e' option of `toplc`.
The idea is to use this option from the `infer.Compile` class, such
that the wrapper can be easily used with Infer as well. So, this is
work towards #52.